Lista 5 - concluida

+++ lista-05/app/src/main/java/engtelecom/poo/LandVehicle.java
@@ -1,7 +1,7 @@
 package engtelecom.poo;
 
 public interface LandVehicle {
-    public abstract void brake(int i);
+    public abstract void brake(double i);
 
-    public abstract void speedUp(int i);
+    public abstract void speedUp(double i);
 }
